agent_core/pyproject.toml
[tool.poetry]
name = "ai-agent-core"
version = "1.0.0"
description = "AI Agent 核心服务 - 基于 LangGraph 和 Instructor 的智能任务处理引擎"
authors = ["AI App Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.12"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
openai = "^1.84.0"
anthropic = "^0.52.2"
google-genai = "^1.18.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
redis = "^5.0.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 